Move enforcement API code to oso.py
