+++ languages/python/oso/oso/__init__.py
@@ -1,5 +1,4 @@
 from polar import polar_class, Variable, Predicate
-from .oso import Oso, Policy
+from .oso import Oso
 from .exceptions import AuthorizationError, ForbiddenError, NotFoundError
 from polar.exceptions import OsoError
-from .enforcer import Enforcer
